feat: 合并后端 API + Railway 部署配置
- 合并算法服务到后端 API (单服务架构)
- 添加 AI 路由 (/api, /api/init, /api/move, /ws)
- 支持 DATABASE_URL 环境变量 (Railway PostgreSQL)
- 修复 duration_seconds 为 float 类型
- 添加 WebSocket 独立路由 /ws
- 更新 railway.json 构建配置
- 更新 vite.config.ts 代理配置 (统一 8000 端口)
- 添加 Playwright 测试脚本 (scripts/)
- 清理假数据测试记录

测试:
- HTTP 模式 AI 满分 97 分 ✅
- 数据保存到数据库 ✅
- 排行榜显示正常 ✅

+++ backend/app/schemas/game.py
@@ -11,7 +11,7 @@ class GameCreate(BaseModel):
     map_size: int = Field(..., ge=5, description="地图尺寸")
     mode: str = Field(..., pattern="^(ai|human)$", description="游戏模式")
     connection: str = Field(..., pattern="^(ws|http)$", description="连接方式")
-    duration_seconds: int = Field(..., ge=0, description="游戏时长（秒）")
+    duration_seconds: float = Field(..., ge=0, description="游戏时长（秒）")
     created_at: Optional[datetime] = None
 
 
@@ -24,7 +24,7 @@ class GameResponse(BaseModel):
     map_size: int
     mode: str
     connection: str
-    duration_seconds: int
+    duration_seconds: float
     created_at: datetime
 
     class Config:
